Replace reconstruction structs with hydro structs
I moved the reconstruction::Primitive struct into the hydro utilities.
Along with that I added a struct for conserved variables and a small
struct for vector quantities that just has members x, y, and z to make
storage and movement of vectors easier. I also made the primitive and
conserved structs use that vector struct for their vector quantites.

Once that was done I fixed all the references to
reconstructon::Primitive to point at the new location/version of the
struct, updated it's usage, and updated all the relevant tests.